Migrate to v2.0 and correct examples
