+++ library.properties
@@ -1,0 +1,10 @@
+name=spark-dallas-temperature
+version=0.0.6
+license=MIT
+author=Tom de Boer <tom@tomdeboer.nl>
+sentence=A Spark Core ported, but otherwise untouched libary for controlling Dallas Temperature Sensors (DS18B20, DS18S20, DS1822,DS1820)
+# paragraph=a longer description of this library, always prepended with sentence when shown
+url=https://github.com/tomdeboer/SparkCoreDallasTemperature
+repository=https://github.com/tomdeboer/SparkCoreDallasTemperature.git
+# architectures=a list of supported boards if this library is hardware dependent, like particle-photon,particle-electron
+dependencies.OneWire=2.0.1
