+++ src/spark-dallas-temperature/spark-dallas-temperature.h
@@ -1,0 +1,1 @@
+#include "../spark-dallas-temperature.h"
